+++ lib/src/core/anonymous_service.dart
@@ -5,17 +5,17 @@ import 'service.dart';
 ///
 /// Well-suited for testing.
 class AnonymousService extends Service {
-  Future Function([Map]) _index;
-  Future Function(Object, [Map]) _read, _create, _remove;
-  Future Function(Object, Object, [Map]) _modify, _update;
+  FutureOr Function([Map]) _index;
+  FutureOr Function(Object, [Map]) _read, _create, _remove;
+  FutureOr Function(Object, Object, [Map]) _modify, _update;
 
   AnonymousService(
-      {Future index([Map params]),
-      Future read(id, [Map params]),
-      Future create(data, [Map params]),
-      Future modify(id, data, [Map params]),
-      Future update(id, data, [Map params]),
-      Future remove(id, [Map params])})
+      {FutureOr index([Map params]),
+      FutureOr read(id, [Map params]),
+      FutureOr create(data, [Map params]),
+      FutureOr modify(id, data, [Map params]),
+      FutureOr update(id, data, [Map params]),
+      FutureOr remove(id, [Map params])})
       : super() {
     _index = index;
     _read = read;
@@ -26,27 +26,28 @@ class AnonymousService extends Service {
   }
 
   @override
-  index([Map params]) => _index != null ? _index(params) : super.index(params);
+  index([Map params]) => new Future.sync(
+      () => _index != null ? _index(params) : super.index(params));
 
   @override
-  read(id, [Map params]) =>
-      _read != null ? _read(id, params) : super.read(id, params);
+  read(id, [Map params]) => new Future.sync(
+      () => _read != null ? _read(id, params) : super.read(id, params));
 
   @override
-  create(data, [Map params]) =>
-      _create != null ? _create(data, params) : super.create(data, params);
+  create(data, [Map params]) => new Future.sync(() =>
+      _create != null ? _create(data, params) : super.create(data, params));
 
   @override
-  modify(id, data, [Map params]) => _modify != null
+  modify(id, data, [Map params]) => new Future.sync(() => _modify != null
       ? _modify(id, data, params)
-      : super.modify(id, data, params);
+      : super.modify(id, data, params));
 
   @override
-  update(id, data, [Map params]) => _update != null
+  update(id, data, [Map params]) => new Future.sync(() => _update != null
       ? _update(id, data, params)
-      : super.update(id, data, params);
+      : super.update(id, data, params));
 
   @override
-  remove(id, [Map params]) =>
-      _remove != null ? _remove(id, params) : super.remove(id, params);
+  remove(id, [Map params]) => new Future.sync(
+      () => _remove != null ? _remove(id, params) : super.remove(id, params));
 }