+++ test/all.dart
@@ -9,6 +9,7 @@ import 'di_test.dart' as di;
 import 'encoders_buffer_test.dart' as encoders_buffer;
 import 'exception_test.dart' as exception;
 import 'extension_test.dart' as extension;
+import 'find_one_test.dart' as find_one;
 import 'general_test.dart' as general;
 import 'hooked_test.dart' as hooked;
 import 'parameter_meta_test.dart' as parameter_meta;
@@ -34,6 +35,7 @@ main() {
   group('encoders_buffer', encoders_buffer.main);
   group('exception', exception.main);
   group('extension', extension.main);
+  group('find_one', find_one.main);
   group('general', general.main);
   group('hooked', hooked.main);
   group('parameter_meta', parameter_meta.main);